Removed the mixin
